levers and loaders: the small batch from the upstream audit
Thirteen items that were each too small to be worth a run of their own, and
together are worth one. All of them come out of the upstream study; none of
them needed hardware or retail assets, which is what put them in one batch.

Loaders, where the study predicted everything unadopted would be:

  R_inPVS read one byte before the start of the PVS row when either point was
  inside a solid leaf. cluster is -1 there, and -1 >> 3 is -1. Both points now
  carry the guard, not just one: the answer is meant to be the same whichever
  way round the caller passes them.

  novis was allocated and filled eight times larger than anything reads. The
  rounding to 64 is in bits and the buffer is in bytes.

  MD3 LOD registration had two defects, in two copies of the same function.
  The loop counted down, stopped at the first failure, and then filled the gaps
  by duplicating mod->data.mdv[lod + 1] - the very entry that had just failed to
  load. It now counts up and the slot is mod->numLods, so every slot below it
  holds something real; R_ComputeLOD already clamps, so nothing needs the
  duplicates. And the extension was cut with strchr, which finds the first dot
  in the whole path, so any directory with a dot in its name broke the load.

A mine under the section of the backlog that raises light limits: the clamp on
num_dlights was written as the literal 32 rather than MAX_DLIGHTS. Correct
today, and it would have survived a raise of the constant silently.

Levers that had been nailed shut. r_vbo is a cvar again rather than three
unconditional qtrue assignments, because "turn it off and try again" is how a
defect on a machine we do not have gets halved. r_neatsky drops mip and picmip
on skybox faces, which is where the seam between faces comes from. r_debugStyle
lights one lightmap style and blacks the rest. r_printShaders dumps what the
material files actually define.

Two knobs that upstream also has and that we deliberately did not copy as-is:

  r_ghoul2timebase is registered by rd-vanilla and never read by it, so the
  knob was dead there. Porting it dead would have been worse than not porting
  it. It now selects the clock, and its default is what this code already did.

  r_fbo is not added at all. The only branch that would render straight to the
  swapchain lives inside an "#if 0", so the cvar would accept a value and not
  do what its name says.

Two cvars had drifted from the spelling the rest of the Quake 3 family uses,
so a config written for any other engine did nothing here and said nothing
about it - and check_cvars.py cannot catch that, because it looks for two
spellings of one name and these were two different names. Renamed, with the
old name read once at startup so nobody's config stops working silently.

Types: const restored on RB_AddQuadStamp and RB_AddQuadStampExt, qboolean on
R_MergedWidthPoints and R_MergedHeightPoints. R_InitFogTable says powf, which
is what pow already resolved to - measured, not assumed: the object code is
byte-identical and both emit a call to powf. Not the same case as the atan2
change, where the file moved from C to C++ and the resolution really did change
under an unchanged source line.

Full local run, all stages green.

+++ code/rd-vulkan/tr_image.cpp
@@ -243,7 +243,13 @@ void R_InitFogTable( void ) {
 	exp = 0.5;
 
 	for ( i = 0 ; i < FOG_TABLE_SIZE ; i++ ) {
-		d = pow ( (float)i/(FOG_TABLE_SIZE-1), exp );
+		// powf, not pow. Measured, not assumed: this file is C++, so pow(float,
+		// float) already resolves to the float overload - the object code is
+		// byte-identical and both spellings emit a call to powf. Upstream writes
+		// powf; matching it means the reader does not have to know that rule.
+		// Not the same case as section 40.1, where the file moved from C to C++
+		// and the resolution really did change under an unchanged source line.
+		d = powf( (float)i/(FOG_TABLE_SIZE-1), exp );
 
 		tr.fogTable[i] = d;
 	}
